requirements.txt
alembic==1.7.6
aniso8601==9.0.1
appdirs==1.4.4
astroid==2.9.3
atomicwrites==1.4.0
attrs==21.4.0
baron==0.10.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==7.1.2
colorama==0.4.4
cors==1.0.1
cryptography==36.0.1
docutils==0.16
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.1.6
Markdown==3.1.1
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.12.0
oauthlib==3.2.0
packaging==21.3
parso==0.6.0
platformdirs==2.5.0
pluggy==0.13.1
py==1.11.0
pycparser==2.21
pylint==2.12.2
PyMySQL==1.0.2
pyparsing==3.0.7
PySocks==1.7.1
pytest==5.3.5
pytest-json-report==1.2.1
pytest-metadata==1.8.0
pytest-sugar==0.9.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.3
redbaron==0.9.2
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rply==0.7.8
six==1.16.0
SQLAlchemy==1.4.31
termcolor==1.1.0
tldextract==3.1.2
toml==0.10.2
tweepy==4.5.0
typer==0.0.8
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
